[general] added  abstract models
